wip:fees

+++ src/interfaces/IkStakingVault.sol
@@ -9,8 +9,8 @@ interface IkStakingVault {
                         USER STAKING OPERATIONS
     //////////////////////////////////////////////////////////////*/
 
-    function requestStake(address to, uint96 kTokensAmount) external payable returns (uint256 requestId);
-    function requestUnstake(address to, uint96 stkTokenAmount) external payable returns (uint256 requestId);
+    function requestStake(address to, uint256 kTokensAmount) external payable returns (uint256 requestId);
+    function requestUnstake(address to, uint256 stkTokenAmount) external payable returns (uint256 requestId);
     function claimStakedShares(bytes32 batchId, uint256 requestIndex) external payable;
     function claimUnstakedAssets(bytes32 batchId, uint256 requestIndex) external payable;
     function updateLastTotalAssets(uint256 totalAssets) external;
